add savedItems, NoItemsView withAnimation

+++ TodoList/Views/ListView.swift
@@ -12,22 +12,31 @@ struct ListView: View {
     @EnvironmentObject var listViewMode: ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewMode.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewMode.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewMode.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List{
+                    ForEach(listViewMode.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewMode.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete { IndexSet in
+                        listViewMode.deleteItem(indexSet: IndexSet)
+                    }
+                    .onMove { IndexSet, newIndex in
+                        listViewMode.moveItem(from: IndexSet, to: newIndex)
+                    }
+                }
             }
-            .onDelete { IndexSet in
-                listViewMode.deleteItem(indexSet: IndexSet)
-            }
-            .onMove { IndexSet, newIndex in
-                listViewMode.moveItem(from: IndexSet, to: newIndex)
-            }
+            
         }
+        
         .listStyle(PlainListStyle())
         .navigationTitle("Todo List 🗒️")
         .navigationBarItems(
